[objc-gnustep2] Protocol cleanups.
Don't emit protocols unless they are referenced.  This should give quite
large code-size savings, because for other GNU ABIs we were emitting
protocols in every compilation unit and relying on the runtime to
deduplicate them.

Emit protocol references via an indirection layer.  This allows the
runtime to change the representation of protocols (and to deduplicate
them and make pointer comparison work).